Force conditional decodablilty to be explicit
