+++ ConfCore/DateFormatter+ConfCoreFormat.swift
