+++ ConfCore/Keyword.swift
@@ -25,12 +25,9 @@ public class Keyword: Object, Decodable {
     // MARK: Decodable
 
     public convenience required init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        let name = try container.decode(String.self)
-
         self.init()
 
-        self.name = name
+        self.name = try decoder.singleValueContainer().decode()
     }
 
 }
